+++ docs/src_examples/cosmo_model/5.py
@@ -4,7 +4,7 @@ import numpy as np
 redshifts = np.linspace(0, 2.5, 40)
 
 theory = model.theory["classy"]
-theory.needs(Hubble={"z": redshifts})
+theory.must_provide(Hubble={"z": redshifts})
 
 omega_cdm = [0.10, 0.11, 0.12, 0.13, 0.14]